rename tsx filenames to camelcase
