+++ app/packages/looker-3d/src/action-bar/PointSize.tsx
