+++ app/packages/looker-3d/src/action-bar/index.tsx
@@ -1,6 +1,6 @@
-export * from "./color-space";
-export * from "./point-size";
-export * from "./screenshot";
-export * from "./set-view";
-export * from "./view-help";
-export * from "./view-json";
+export * from "./ColorSpace";
+export * from "./PointSize";
+export * from "./Screenshot";
+export * from "./SetViewButton";
+export * from "./ViewHelp";
+export * from "./ViewJson";
